feat: adding mobile view
