+++ pubspec.yaml
@@ -27,6 +27,7 @@ dependencies:
 
   authentication_repository:
     path: packages/authentication_repository
+  responsive_builder: ^0.4.1
 
 dev_dependencies:
   flutter_test:
